+++ Include/ramio/database/database.h
@@ -1,1 +1,1 @@
-#include "../../../Source/Database/Database.h"
+#include "../../../Source/Database/DatabaseConnection.h"